[No-ticket] Using Results<T> instead of Exceptions (#61)
Background
----------

In some cases, we're using exceptions to communicate to controllers the result of operations. In some other cases, we don't have a good mechanism of indicating errors that preventing the operation from executing without exceptions at all.

There is a pattern called the Maybe Monad, the Either Monad, or the Left/Right monad that all have the same concept to solve this. We're introducing it here as `Result<T>`. More on this below.

Why we don't want to use exceptions:

1. Exceptions are semantically meant to only capture "exceptions" in code flow. **Errors are not exceptions.** Errors are controlled, handled, and they cover part of the functionality that the system is supposed to cover. For example, a user not being found is not an "exceptional" situation, but rather a situation that the system should account for. As the system matures, more exceptions are controlled and the system has fewer exceptions and more controlled error situations.

2. Exceptions have a minor performance impact. Every time we `try` for an exception, we are creating a context in which exceptions can be handle. This impacts memory and speed since the stack and the progression of the flow need to be accounted for to properly report when an exception happens.

3. It is common to miss information between catching and throwing exceptions. When an exception is thrown, that very same exception should be reported to our metric systems, not a wrapped exception. This makes it easier to troubleshoot the real flow of that exception.

4. Alarm systems (that we're currently working on building) will report exceptions as problems to be fixed. Fake-alarms will work against us. tried/caught exceptions will also be reported, because even when not crashing the system, they are still issues to be solved (see point 1.) This will be also important for our APM reporting.

5. Creating try/catch around the code is just bothersome. It forces an unnatural flow of code that makes it difficult to follow, especially when trying to recover from an exception, which creates unexpected flows (try/catch/finally with local values, nested try/catch).

<h3>The Result<T> Object</h3>

In short, the `Result<T>` object is a wrapper around `T` that allows us to check for the result of an operation that was supposed to return `T`. This way, we can communicate the result without throwing exceptions, and treating them appropriately.

We can report a failed operation by using `Result.Fail()` which returns a typified error. This is similar to how people use exceptions, but without the actual exception. If you're feeling fancy, you can add multiple errors in `Reasons` for an error, which can be as complex as you want.

We can result in successful operations by using `Result.Ok()`, or even `Result.OkIf()` if you're feeling fancy.

We can check on the result by using `result.IsFailed`,  `report.IsSuccess`, and we can even check for specific errors with `result.HasError<TError>`.

For more information, check out the NuGet library that powers this: [FluentResults](https://github.com/altmann/FluentResults).

Changes done
-----

- Implemented these changes in a few examples. Did not change any of the underlying behavior.
- Created a helper that can transform any Result known error into a meaningful API Response.

Pending changes
-------

- Review the rest of the operations and apply this pattern if you agree to follow it.
- Review the pending TODOs that I've marked, for issues detected, and fix at some point.
- Review the pending unit tests which are not included in the repository (02 Domain folder is completely missing).
- Decide if we want meaningful API Responses to include details about the error. Since the errors come from internal layers of the app, I have opted not to do so.
- I applied most of this logic in the Domain layer (Services). Consider doing something similar for the Data layer.